+++ src/routes/+page.svelte
@@ -22,6 +22,7 @@
 	import SignOut from '@icons/SignOut.svelte';
 	import { oneTime, rateClickLimiter } from '@utils/general';
 	import { error, info, success } from '@utils/notifier';
+	import Header from '@Layouts/Header.svelte';
 
 	let dialog: Dialog;
 
@@ -176,15 +177,8 @@
 	<div
 		class="flex flex-col lg:flex-row h-screen justify-around items-center bg-background-reverse snap-start relative"
 	>
-		<DropdownMenu
-			class="absolute ltr:left-0 rtl:right-0 top-0 p-3"
-			bind:open
-			profile={{
-				name: $LL.DropDown.Name(),
-				description: $LL.DropDown.Description(),
-				href: '#',
-				imgSrc: 'https://i.pinimg.com/originals/25/bd/8b/25bd8b7f6e57cdfd17747b25d753b2ce.jpg'
-			}}
+		<Header
+			imgSrc="https://i.pinimg.com/originals/25/bd/8b/25bd8b7f6e57cdfd17747b25d753b2ce.jpg"
 			items={[
 				{ id: 0, text: $LL.DropDown.ViewProfile(), href: '#', icon: ViewProfile },
 				{ id: 1, text: $LL.DropDown.Settings(), href: '#', icon: Settings, lineBelow: true },
@@ -193,19 +187,41 @@
 				{ id: 4, text: $LL.DropDown.Help(), href: '#', icon: Help, lineBelow: true },
 				{ id: 5, text: $LL.DropDown.SignOut(), href: '#', icon: SignOut }
 			]}
-			><span class="mx-1">{$LL.DropDown.Name()}</span>
-			<svg
-				data-open={open}
-				class="w-5 h-5 mx-1 data-[open=true]:rotate-180"
-				viewBox="0 0 24 24"
-				fill="none"
-				xmlns="http://www.w3.org/2000/svg"
-			>
-				<path
-					d="M12 15.713L18.01 9.70299L16.597 8.28799L12 12.888L7.40399 8.28799L5.98999 9.70199L12 15.713Z"
-					fill="currentColor"
-				></path>
-			</svg></DropdownMenu
-		>
+		/>
+		<div class="h-full w-full container lg:pt-[76px] sm:pt-[60px] pt-[49px]">
+			<div class="h-full relative">
+				<DropdownMenu
+					bind:open
+					class="absolute ltr:left-0 rtl:right-0 top-0 p-3"
+					profile={{
+						name: $LL.DropDown.Name(),
+						description: $LL.DropDown.Description(),
+						href: '#',
+						imgSrc: 'https://i.pinimg.com/originals/25/bd/8b/25bd8b7f6e57cdfd17747b25d753b2ce.jpg'
+					}}
+					items={[
+						{ id: 0, text: $LL.DropDown.ViewProfile(), href: '#', icon: ViewProfile },
+						{ id: 1, text: $LL.DropDown.Settings(), href: '#', icon: Settings, lineBelow: true },
+						{ id: 2, text: $LL.DropDown.Team(), href: '#', icon: Team },
+						{ id: 3, text: $LL.DropDown.InviteColleagues(), href: '#', icon: InviteColleagues },
+						{ id: 4, text: $LL.DropDown.Help(), href: '#', icon: Help, lineBelow: true },
+						{ id: 5, text: $LL.DropDown.SignOut(), href: '#', icon: SignOut }
+					]}
+					><span class="mx-1">{$LL.DropDown.Name()}</span>
+					<svg
+						data-open={open}
+						class="w-5 h-5 mx-1 data-[open=true]:rotate-180"
+						viewBox="0 0 24 24"
+						fill="none"
+						xmlns="http://www.w3.org/2000/svg"
+					>
+						<path
+							d="M12 15.713L18.01 9.70299L16.597 8.28799L12 12.888L7.40399 8.28799L5.98999 9.70199L12 15.713Z"
+							fill="currentColor"
+						></path>
+					</svg></DropdownMenu
+				>
+			</div>
+		</div>
 	</div>
 </div>